Apply a universal 1.5x playback speed to computer-use recordings
Wires the server-owned playback speed multiplier forward to both
substrates instead of only macOS, and switches macOS's live
uniform-speed default from 4x to 1.5x to match.

- crates/ai: read the new fractional playback_speed wire field
  (crates/ai/src/agent/action/convert.rs); widen
  AIAgentActionType::StartRecording::playback_speed_multiplier from
  Option<u32> to Option<f32>. f32 isn't Eq, so AIAgentActionType and
  every type that transitively wraps it (AIAgentAction,
  AIAgentOutputMessageType, AIAgentOutputMessage in the warp crate;
  TuiAIBlockSection in warp_tui) drop the Eq derive and keep
  PartialEq, which is sufficient everywhere they're used.
- crates/computer_use: bump the client fallback default from 4.0 to
  1.5 (crates/computer_use/src/lib.rs). macOS live setpts filter
  already supported fractional values, so no mac/recording.rs change
  is needed beyond the default and tests. Linux gets a new final
  ffmpeg speed pass (apply_playback_speed in
  crates/computer_use/src/linux/recording.rs) applied strictly after
  the smart cut and overlay burn-in, so QUALITY-1112 overlay remap
  and 1x-master invariant are preserved; the live x11grab capture
  stays unaffected.
- app/src/ai: resolve the fractional multiplier in
  execute/start_recording.rs, thread it through ActiveRecording
  (recording_controller.rs) into recording_finalize.rs call to
  post_process_recording.
- Updated the QUALITY-1112 spec invariant 11 and Playback multiplier
  resolution to describe the new final pass instead of the old
  ignored-on-Linux behavior.
- Wired up crates/ai/src/agent/action/convert_tests.rs into the
  module tree (it existed but was never included via a path mod
  declaration, so it never actually compiled or ran) and added
  coverage for fractional/real-time playback_speed conversion.
- Bumped the warp_multi_agent_api git dependency to the
  warp-proto-apis feature branch commit that adds the new
  playback_speed field. This must be re-pinned to a commit on
  warp-proto-apis main once that PR merges, before this PR merges.

+++ crates/ai/src/agent/action/mod.rs
@@ -33,7 +33,10 @@ use crate::diff_validation::ParsedDiff;
 use crate::document::AIDocumentId;
 use crate::skills::SkillReference;
 
-#[derive(Debug, Clone, Eq, PartialEq, EnumDiscriminants)]
+// Not `Eq`: `StartRecording::playback_speed_multiplier` is an `Option<f32>`,
+// and `f32` does not implement `Eq` (NaN is not reflexive). `PartialEq` is
+// sufficient for every current use of this type.
+#[derive(Debug, Clone, PartialEq, EnumDiscriminants)]
 pub enum AIAgentActionType {
     /// The AI requested the output for a given command to be retrieved as context in responding to
     /// a user's query.
@@ -147,15 +150,16 @@ pub enum AIAgentActionType {
     /// arrives on the tool call; the client applies it. `frame_rate` of 0 means
     /// unset. `summary` is a short agent-authored title shown in badges.
     /// `description` is an optional longer description shown in detail views.
-    /// `playback_speed_multiplier` is the integer speed factor from the proto
-    /// (e.g. 4 = 4×). `None` or a value ≤ 1 means real-time (use client default).
+    /// `playback_speed_multiplier` is the fractional speed factor from the
+    /// proto's `playback_speed` field (e.g. 1.5 = 1.5×). `None` or a value ≤ 1
+    /// means real-time (use client default).
     StartRecording {
         frame_rate: u32,
         max_duration: Option<Duration>,
         max_size_bytes: Option<u64>,
         summary: Option<String>,
         description: Option<String>,
-        playback_speed_multiplier: Option<u32>,
+        playback_speed_multiplier: Option<f32>,
         /// The surface to record. `None` records the whole screen; a `Window`
         /// target records just that window via native ffmpeg `x11grab
         /// -window_id` on the foreground-visible window. Applied by the client
